# swagger.yaml
openapi: 3.0.0
info:
  title: E Commerce App 2
  description: Codecademy Backend Portfolio project.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server for studying...
paths:
...
  /users/{userId}:
    get:
      summary: Return a user data by userId.
      description: Return a user data by userId.
      parameters:
        - name: userId
          in: path
          required: true
          description: userId for specify an existing user.
          schema:
            type: integer
            minimum: 1
      responses:
        "200": # status code
          description: A JSON object of user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      summary: Update a user data by userId.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: A JSON object of a user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 123
        name:
          type: string
          example: Jon Doe
        email:
          type: string
          format: email
          example: jondoe@example.com
        password:
          type: string
          format: password
          example: password